build: [#42309] Runner with Kubernetes 2
This changes removes the explicit tags that sometimes makes this job to
run on the deprecated Gitlab Runner with the Docker executor and the
`high-mem`-tag. Without this, the jobs will run runners with the "Run
untagged jobs" configuration. Currently the only runner with this
setting is the Gitlab Runner with the Kubernetes executor. It have more
resource that the docker runner with the `high-mem`-tag.